Code reorganisation and cleanup.
- Move demography stuff to demography.py
- Move utils->core and make this first import.
- Rename 'simulation.py' to 'ancestry.py'
- Removed lots of dead test code.
